Unit tests for Push LocalStorageManager
